test: add editor-tools tests
